Working interface with PyMC3, but shit
Sampling seems to work, although it doesn't actually calculate
the likelihood at the end like it's supposed to so it seems
like I would have to re-write everything in Theano which is not
something I feel a burning desire to do.

For now I think I'll implement a `emcee` backend, just like how the DR1
stuff was done.