Added more interfacing to `EmceeHelper`
Now added `posterior` property to helper class, which returns an arviz
`InferenceData` object. Accompanying this is a wrapper function to dump
the posterior samples to a NetCDF file.

Started working on a convenience function to convert the posterior
samples into formatted prior model for subsequent usage. This is not
finished.