Added a nominal VLSR offset prior to trim_array
Catalog frequencies were being chopped off, and became apparent when
working on HC9N. A nominal VLSR is included as part of the model, which
offsets the spectrum frequencies to the source frame. This value should
correspond to the same value used during `preprocess_spectrum` in the
data cleaning step.