Implementing mechanics for parameter freeze
In Base, there is now a DeltaLikelihood and DeltaParameter pair of
classes for freezing parameters; a value is provided, and is infinitely
unlikely in all parameter space except that value.

Associated with this is changing the way that initial parameters and
"scaling" is performed. Previously, an array of values is tiled over the
number of walkers, and every parameter is shifted by a small amount to
start the walkers off, controlled by the `scale` kwarg.

Now, `scale` can be specified as `None` for backwards compatibility. If
`scale=None`, then it uses the new `AbstractModel` method `sample_prior`
for initializing the walker positions. This actually samples values from
the defined distributions, so that the `DeltaLikelihood` is also
supported (it just returns the set value no matter what).