Changed restart mode for sampling
Previously we would use the resample the posterior to restart runs. This
is probably not what we actually want to do, as a restart should
actually be using the last step of every walker instead.

The likelihood check should also be based on the mean of either the
prior (for fresh starts with no scale) or the walker average of the last
step (for restarts).