Added caching of indexes for Issue #4
In reference to issue #4, I've modified the behaviour of `_make_lines`
in the `Simulation` class. In the past, every iteration will
re-calculate the indexes to use for the line profile simulation, which
was contributing more or less half of the runtime during an MCMC run.

The routine will now do a check for the `use_obs` flag, which is
primarily designed for MCMC analysis. If this is set to `True`, we check
the cache to see if the indexes have already been calculated; if so, we
simply reuse the cached indexes and if not run `find_nearest` on the
data.

I checked this on HC9N, and does not appear to affect the MCMC result.