Added log sampling behavior for column densities
The `SingleComponent` model, which forms the basis for the more complex
models, will now assume we are sampling in log space when `ncol` is
sufficiently small upon when `simulate_spectrum` is called.

I'm hoping that sampling in log space will help improve convergence a
bit, as the scale of moves being proposed is not large enough to match
the large values that column densities usually have. I have also
included a test routine to make sure that the likelihood calculation
matches.