Implemented a cospatial model
The `CospatialTMC1` inherits from the `TMC1FourComponent` model,
primarily to include the constraints placed on the four velocity
components. The main behavioral difference is only one source size is
used in all of the likelihood and spectral simulation backends.

A test has also been devised to ensure the model can be created, and
that the methods can run without a hitch.