Refactoring prior constraints
Previously, the TMC1FourComponent model contained the velocity
constraint within the `compute_prior_likelihood` function override.
Since that wasn't a great organization of code, this constraint is now
moved to the `prior_constraint` function with more or less the same
code, just moved in here for organization.

The idea is to minimize edits to the likelihood calculation as to
mitigate potential errors, and instead edit the `prior_constaint`
function.