Added a bias term for fine control of masking
The `SpectrumChunk` mask setter method now also takes an additional bias
term in the `parameters` tuple. This value is used to shift the
baseline, so to speak, for signal detection and masking. Basically, it
allows you to shift the threshold up and down, depending on if you want
to mask everything or nothing.

This functionality is primarily used for simulation masking, where the
RMS may not necessarily be exactly zero, and somewhat hard to program a
heuristic for. Instead, when masks are being set, we just artificially
shift the threshold to negative numbers, and since the ROI is protected,
we just mask everything in the simulation except the ROI. This works
well for the methyl chains.