Finalized the new velocity stacking routine
Added quite a lot of docstrings to the `stack` module, as well as
confirming that the `velocity_stack_pipeline` is now working as
intended, intelligently masking regions where we know there will be
lines from the same molecule we are stacking, offset in velocity. The
main case for this are instances where lines are very close to one
another, such as the case of prolate top K-ladders.